Databases/Project/requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
psycopg2-binary